# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: PU Transit API
servers:
  - url: /api
    description: Base API path
security:
  - FirebaseBearer: []
...
paths:
...
  /routes:
    get:
      operationId: listRoutes
      tags: [routes]
      summary: List routes
      responses:
        "200":
          description: Routes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Route"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
    post:
      operationId: saveRoute
      tags: [routes]
      summary: Create or update a route
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RouteInput"
      responses:
        "200":
          description: Updated route
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Route"
        "201":
          description: Created route
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Route"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /routes/{id}:
    put:
      operationId: putRoute
      tags: [routes]
      summary: Create or replace a route under a client-chosen UUID (RTE-04) — idempotent, so a save whose response was lost can be retried without creating a second route; non-UUID ids are 400
      parameters:
        - $ref: "#/components/parameters/RouteId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RouteInput"
      responses:
        "200":
          description: Replaced route
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Route"
        "201":
          description: Created route
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Route"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
    patch:
      operationId: updateRoute
      tags: [routes]
      summary: Update a route by ID
      parameters:
        - $ref: "#/components/parameters/RouteId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RouteInput"
      responses:
        "200":
          description: Updated route
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Route"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
    delete:
      operationId: deleteRoute
      tags: [routes]
      summary: Permanently delete a never-published draft (RTE-03); published routes are archived instead (409 ROUTE_PUBLISHED)
      parameters:
        - $ref: "#/components/parameters/RouteId"
      responses:
        "204":
          description: Deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
...
  /assignments:
    get:
      operationId: listAssignments
      tags: [assignments]
      summary: List dated assignments, soonest first (admin)
      responses:
        "200":
          description: Assignments sorted by service date, shift and bus
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Assignment"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
    post:
      operationId: createAssignment
      tags: [assignments]
      summary: Put a driver on a bus and its published route for one service date (ASG-01); overrides the standing bus for that date and shift
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignmentInput"
      responses:
        "201":
          description: Saved assignment (the same driver, bus, date and shift overwrites its route)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assignment"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          description: ASSIGNMENT_CONFLICT names the other driver or bus for that date and shift; DRIVER_NOT_ELIGIBLE, BUS_NOT_REGISTERED, BUS_OUT_OF_SERVICE or ROUTE_NOT_FOR_BUS explain a refused reference
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /assignments/{driverUid}/{assignmentId}:
    delete:
      operationId: deleteAssignment
      tags: [assignments]
      summary: Remove a dated assignment (admin); refused while a live trip runs under it (409 ASSIGNMENT_IN_USE)
      parameters:
        - name: driverUid
          in: path
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 128
        - name: assignmentId
          in: path
          required: true
          schema:
            type: string
            pattern: "^\\d{4}-\\d{2}-\\d{2}_[a-z0-9]{1,40}_[A-Z0-9]{2,20}$"
      responses:
        "204":
          description: Removed (or already absent)
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
...
components:
  securitySchemes:
    FirebaseBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    RouteId:
      name: id
      in: path
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 128
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: Tracking ownership or replay conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Missing or invalid Firebase JWT
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Firebase email, membership, role, or assignment is not allowed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    ServiceUnavailable:
      description: Firebase or migration service unavailable
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    Error:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: string
        code:
          type: string
...
    RouteStop:
      type: object
      additionalProperties: false
      required: [lat, lng]
      properties:
        lat:
          type: number
          minimum: -90
          maximum: 90
        lng:
          type: number
          minimum: -180
          maximum: 180
        name:
          type: string
          maxLength: 200
    Coordinate:
      type: object
      additionalProperties: false
      required: [lat, lng]
      properties:
        lat:
          type: number
          minimum: -90
          maximum: 90
        lng:
          type: number
          minimum: -180
          maximum: 180
...
    Route:
      type: object
      additionalProperties: false
      required: [id, shift, busNumber, origin, destination, stops, status]
      properties:
        id:
          type: string
        status:
          $ref: "#/components/schemas/RouteStatus"
        publishedVersion:
          description: Number of the version riders currently see (RTE-02); absent until the route is first published under versioning
          type: integer
          minimum: 1
        hasDraft:
          description: Admin lists only — the returned content is an unpublished draft; riders still see publishedVersion
          type: boolean
        pathSource:
          $ref: "#/components/schemas/RoutePathSource"
        shift:
          type: string
        busNumber:
          type: string
        origin:
          type: string
        destination:
          type: string
        stops:
          type: array
          maxItems: 500
          items:
            $ref: "#/components/schemas/RouteStop"
        pathData:
          type: array
          maxItems: 5000
          items:
            $ref: "#/components/schemas/Coordinate"
        kind:
          $ref: "#/components/schemas/RouteKind"
    RouteKind:
      description: Regular city bus or a campus shuttle; absent means bus
      type: string
      enum: [bus, shuttle]
    RouteStatus:
      description: Drafts are visible to admins only; publishing requires a verified path or the manual-path acknowledgement (RTE-01). Records saved before this field existed read as published. Archived routes are hidden from riders and kept for history (RTE-03 adds the action; only draft and published are accepted as input).
      type: string
      enum: [draft, published, archived]
...
    RoutePathSource:
      description: How the drawn path was produced. ors = road geometry from the directions provider through every stop (verified by the API to pass within 150 m of each stop); manual = an admin acknowledged the path is not road-verified and students see "path not verified". Absent = unverified, publishable only as a draft.
      type: string
      enum: [ors, manual]
    RouteInput:
      type: object
      additionalProperties: false
      required: [shift, busNumber, origin, destination, stops, status]
      properties:
        id:
          type: string
          maxLength: 128
        status:
          description: The action - draft or published. Archiving is its own endpoint (RTE-03).
          type: string
          enum: [draft, published]
        pathSource:
          $ref: "#/components/schemas/RoutePathSource"
        shift:
          type: string
          minLength: 1
          maxLength: 200
        busNumber:
          type: string
          minLength: 1
          maxLength: 200
        origin:
          type: string
          minLength: 1
          maxLength: 200
        destination:
          type: string
          minLength: 1
          maxLength: 200
        stops:
          type: array
          maxItems: 500
          items:
            $ref: "#/components/schemas/RouteStop"
        pathData:
          type: array
          maxItems: 5000
          items:
            $ref: "#/components/schemas/Coordinate"
        kind:
          $ref: "#/components/schemas/RouteKind"
...
    Assignment:
      description: A dated override of the standing bus (ASG-01); the key is serviceDate_shiftKey_busId under the driver
      type: object
      additionalProperties: false
      required: [id, driverUid, driverEmail, busId, routeId, routeVersion, shift, serviceDate, startsAt, endsAt, createdAt, createdBy]
      properties:
        id:
          type: string
          pattern: "^\\d{4}-\\d{2}-\\d{2}_[a-z0-9]{1,40}_[A-Z0-9]{2,20}$"
        driverUid:
          type: string
        driverEmail:
          description: Copied from the membership when saved, so conflict messages can name people
          type: string
        busId:
          type: string
          pattern: "^[A-Z0-9]{2,20}$"
        routeId:
          type: string
          format: uuid
        routeVersion:
          description: The route's published version when the assignment was saved; informational, Start pins its own
          type: ["integer", "null"]
        shift:
          description: The route's shift at save time
          type: string
        serviceDate:
          description: Calendar date in Asia/Kolkata
          type: string
          pattern: "^\\d{4}-\\d{2}-\\d{2}$"
        startsAt:
          description: IST midnight opening the service date, epoch milliseconds
          type: number
        endsAt:
          description: startsAt + 24h
          type: number
        createdAt:
          type: number
        createdBy:
          type: string
    AssignmentInput:
      type: object
      additionalProperties: false
      required: [driverUid, busId, routeId, serviceDate]
      properties:
        driverUid:
          type: string
          minLength: 1
          maxLength: 128
          pattern: "^[A-Za-z0-9][A-Za-z0-9:_-]{0,127}$"
        busId:
          description: Registration as written; normalized to the registry key
          type: string
          minLength: 2
          maxLength: 64
        routeId:
          description: A published route of that bus; its shift becomes the assignment's shift
          type: string
          format: uuid
        serviceDate:
          description: Calendar date in Asia/Kolkata, today or later
          type: string
          pattern: "^\\d{4}-\\d{2}-\\d{2}$"
...