# lib/api-zod/src/generated/types/route.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * PU Transit API
 * OpenAPI spec version: 0.1.0
 */
import type { Coordinate } from './coordinate';
import type { RouteKind } from './routeKind';
import type { RoutePathSource } from './routePathSource';
import type { RouteStatus } from './routeStatus';
import type { RouteStop } from './routeStop';

export interface Route {
  id: string;
  status: RouteStatus;
  /**
     * Number of the version riders currently see (RTE-02); absent until the route is first published under versioning
     * @minimum 1
     */
  publishedVersion?: number;
  /** Admin lists only — the returned content is an unpublished draft; riders still see publishedVersion */
  hasDraft?: boolean;
  pathSource?: RoutePathSource;
  shift: string;
  busNumber: string;
  origin: string;
  destination: string;
  /** @maxItems 500 */
  stops: RouteStop[];
  /** @maxItems 5000 */
  pathData?: Coordinate[];
  kind?: RouteKind;
}