# lib/api-zod/src/generated/types/routeKind.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * PU Transit API
 * OpenAPI spec version: 0.1.0
 */

/**
 * Regular city bus or a campus shuttle; absent means bus
 */
export type RouteKind = typeof RouteKind[keyof typeof RouteKind];


export const RouteKind = {
  bus: 'bus',
  shuttle: 'shuttle',
} as const;