# lib/api-zod/src/generated/types/routePathSource.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * PU Transit API
 * OpenAPI spec version: 0.1.0
 */

/**
 * How the drawn path was produced. ors = road geometry from the directions provider through every stop (verified by the API to pass within 150 m of each stop); manual = an admin acknowledged the path is not road-verified and students see "path not verified". Absent = unverified, publishable only as a draft.
 */
export type RoutePathSource = typeof RoutePathSource[keyof typeof RoutePathSource];


export const RoutePathSource = {
  ors: 'ors',
  manual: 'manual',
} as const;