# lib/api-zod/src/generated/types/routeStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * PU Transit API
 * OpenAPI spec version: 0.1.0
 */

/**
 * Drafts are visible to admins only; publishing requires a verified path or the manual-path acknowledgement (RTE-01). Records saved before this field existed read as published. Archived routes are hidden from riders and kept for history (RTE-03 adds the action; only draft and published are accepted as input).
 */
export type RouteStatus = typeof RouteStatus[keyof typeof RouteStatus];


export const RouteStatus = {
  draft: 'draft',
  published: 'published',
  archived: 'archived',
} as const;