# lib/api-zod/src/generated/types/routeStop.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * PU Transit API
 * OpenAPI spec version: 0.1.0
 */

export interface RouteStop {
  /**
     * @minimum -90
     * @maximum 90
     */
  lat: number;
  /**
     * @minimum -180
     * @maximum 180
     */
  lng: number;
  /** @maxLength 200 */
  name?: string;
}